Skip stream logs test on Windows
