+++ base/commands/viridian/viridian_it_windows_test.go
@@ -1,0 +1,5 @@
+package viridian_test
+
+func streamLogs_nonInteractiveTest(t *testing.T) {
+	t.Skipf("This test doesn't run on Windows")
+}
